feat: OptionMenu 파일명 kebab-case로 변경 및 dim 오버레이 추가 (#8)
